requirements.txt
flask
flask-bootstrap
flask-sqlalchemy
flask-wtf
wtforms
email_validator
flask-bcrypt
flask-login
PyJWT
flask-mail



